completed cnn model
